Move texture conveniences.
