+++ src/geovista/cache.py
@@ -23,17 +23,10 @@ from .config import resources
 
 __all__ = [
     "CACHE",
-    "blue_marble",
-    "checkerboard",
     "fetch_coastlines",
-    "natural_earth_1",
-    "natural_earth_hypsometric",
     "pooch_mute",
     "reload_registry",
 ]
-
-# Type aliases.
-TextureLike = str | pv.Texture
 
 #: Base URL for geovista resources.
 BASE_URL: str = "https://github.com/bjlittle/geovista-data/raw/{version}/data/"
@@ -79,86 +72,6 @@ GEOVISTA_POOCH_MUTE: bool = (
 )
 
 
-def _fetch_texture(fname: str, location: bool | None = False) -> TextureLike:
-    """Get the texture resource from the cache.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    fname : str
-        The base file name of the resource, excluding any directory prefix.
-    location : bool, default=False
-        Determine whether the absolute path filename to the texture resource
-        location within the cache is returned, or the actual texture.
-
-    Returns
-    -------
-    str or Texture
-        The PyVista texture filename or the texture.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    resource = CACHE.fetch(f"raster/{fname}")
-    if not location:
-        resource = pv.read_texture(resource)
-    return resource
-
-
-def blue_marble(location: bool | None = False) -> TextureLike:
-    """Get the NASA Blue Marble Next Generation with topography and bathymetry texture.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    location : bool, default=False
-        Determine whether the absolute path filename to the texture resource
-        location within the cache is returned, or the actual texture.
-
-    Returns
-    -------
-    str or Texture
-        The PyVista texture filename or the texture.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    return _fetch_texture("world.topo.bathy.200412.3x5400x2700.jpg", location=location)
-
-
-def checkerboard(location: bool | None = False) -> TextureLike:
-    """Get the UV checker map 4K texture.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    location : bool, default=False
-        Determine whether the absolute path filename to the texture resource
-        location within the cache is returned, or the actual texture.
-
-    Returns
-    -------
-    str or Texture
-        The PyVista texture filename or the texture.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    return _fetch_texture("uv-checker-map-4k.png", location=location)
-
-
 def fetch_coastlines(resolution: str | None = None) -> pv.PolyData:
     """Get the Natural Earth coastlines for the required resolution.
 
@@ -190,65 +103,6 @@ def fetch_coastlines(resolution: str | None = None) -> pv.PolyData:
     resource = CACHE.fetch(f"natural_earth/physical/{fname}.bz2", processor=processor)
 
     return pv.read(resource)
-
-
-def natural_earth_1(location: bool | None = False) -> TextureLike:
-    """Get the 1:50m Natural Earth texture.
-
-    This is the Natural Earth 1 with shaded relief and water texture.
-
-    This resource has been down-sampled to 65% of its original resolution.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    location : bool, default=False
-        Determine whether the absolute path filename to the texture resource
-        location within the cache is returned, or the actual texture.
-
-    Returns
-    -------
-    str or Texture
-        The PyVista texture filename or the texture.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    return _fetch_texture("NE1_50M_SR_W.jpg", location=location)
-
-
-def natural_earth_hypsometric(location: bool | None = False) -> TextureLike:
-    """Get the 1:50m Natural Earth texture.
-
-    This is the Natural Earth cross-blended hypsometric tints with shaded relief and
-    water texture.
-
-    This resource has been down-sampled to 65% of its original resolution.
-
-    If the resource is not already available in the geovista :data:`CACHE`,
-    then it will be downloaded from the :data:`BASE_URL`.
-
-    Parameters
-    ----------
-    location : bool, default=False
-        Determine whether the absolute path filename to the texture resource
-        location within the cache is returned, or the actual texture.
-
-    Returns
-    -------
-    str or Texture
-        The PyVista texture filename or the texture.
-
-    Notes
-    -----
-    .. versionadded:: 0.1.0
-
-    """
-    return _fetch_texture("HYP_50M_SR_W.jpg", location=location)
 
 
 def pooch_mute(silent: bool = True) -> None:
@@ -297,3 +151,4 @@ def reload_registry(fname: str | None = None) -> None:
 
 # configure the pooch cache manager logger verbosity
 pooch_mute(GEOVISTA_POOCH_MUTE)
+
